+++ History.md
@@ -1,5 +1,14 @@
+1.0.0 / 2019-03-09
+==================
+
+  * Modify function list_cronjob
+  * Modify function list_files
+  * Create function exec_to_array
+
 1.0.0 / 2019-03-07
 ==================
 
+  * Create function list_cronjob
+  * Create function list_files
   * Create function exec
   * Start develop Node Terminal